Change directory numbering to two-digit
There was no guarantee that the directories for each chapter were
listed in their intended, numerical order. GitHub, for example,
lists them alphanumerically, resulting in an ordering where
the two-digit chapter directories are not listed at the end.

To prevent this, all the one-digit chapters now have a zero digit
placed at the beginning of its directory name, resulting in a
fixed two-digit numbering system. This ensures that the chapter
directories will always be listed numerically and thus sorted
correctly.